+++ internal/cmd/grpc.go
@@ -9,6 +9,8 @@ import (
 	"sync"
 	"time"
 
+	"go.flipt.io/flipt/internal/server/ofrep"
+
 	otlpRuntime "go.opentelemetry.io/contrib/instrumentation/runtime"
 
 	"go.opentelemetry.io/contrib/propagators/autoprop"
@@ -258,6 +260,7 @@ func NewGRPCServer(
 		evalsrv     = evaluation.New(logger, store)
 		evaldatasrv = evaluationdata.New(logger, store)
 		healthsrv   = health.NewServer()
+		ofrepsrv    = ofrep.New(cfg.Cache)
 	)
 
 	var (
@@ -276,6 +279,7 @@ func NewGRPCServer(
 	skipAuthnIfExcluded(fliptsrv, cfg.Authentication.Exclude.Management)
 	skipAuthnIfExcluded(evalsrv, cfg.Authentication.Exclude.Evaluation)
 	skipAuthnIfExcluded(evaldatasrv, cfg.Authentication.Exclude.Evaluation)
+	skipAuthnIfExcluded(ofrepsrv, cfg.Authentication.Exclude.OFREP)
 
 	var checker audit.EventPairChecker = &audit.NoOpChecker{}
 
@@ -336,6 +340,7 @@ func NewGRPCServer(
 	register.Add(metasrv)
 	register.Add(evalsrv)
 	register.Add(evaldatasrv)
+	register.Add(ofrepsrv)
 
 	// forward internal gRPC logging to zap
 	grpcLogLevel, err := zapcore.ParseLevel(cfg.Log.GRPCLevel)